Modified paragraphs to not break for columns
This is only fully implemented in Opera right now, so the value isn't
apparent yet, but will be good to have in place when support for
multicolumn improves.

Columns can still be too tall if a section is really rambling and not
broken up.  Can address with JS to replace column-width with a column
max width, but these cases should be rare enough that it can be ignored
for now.