Added some column fitting, header breadcrumbs
If the columns would be too tall to fit in the viewport all at once,
they're combined into a single column.

Columns and headers have been centered.

When scrolling down the page, headers are displayed in breadcrumb style
in a floating section at the top, so you know what section context
you're in at all times.