Added reading progress bars to breadcrumbs
The progress bars underline the section titles in the breadcrumbs, and
fill up as you read through a section.  The bottom scroll position of
the window is used to determine progress.

Basic progress bar styling for Chrome is in place, but a polyfill is
probably called for.

The misalignment of the progress bars is a little jarring and should be
fixed.