Added cross browser progress styles, column layout, breadcrumb and progress tweaks
Styles are more comprehensive for the progress bars and now include
working styles for Firefox, and updated styles for other browsers.
Opera's progress bars are only green.  Sorry.

Breadcrumbs now update on window resize, as does scroll progress.
Bottom margins on the breadcrumbs are now half as tall.

Columns that are too short are now consolidated into one column, the
same way that columns that are too tall for the viewport are.

The height of columns considered too tall is now a little shorter to
account for some padding.